Refactor rejoin logic
If a SED can't rejoin the Zigbee network, as happens during a mains power
outage, zigbee_app_utils eventually gives up and waits for a
user_input_indicate() call.  This means that hours or days after power
is restored, switches can still be (temporarily) unresponsive.  Fix
this by periodically initiating a rejoin attempt.

Also, if the node isn't joined to a Zigbee network when the user presses
a button, flash the LED red rather than yellow.

Signed-off-by: Jeff Zigler <zigboi@nrfhome.com>