zigbee-switch: Recover more quickly after a power failure
If the SED switch loses contact with routers and falls off the network
due to a mains power failure, and then the user presses a button while
it's in that state, force a chip reset so that the switch can bypass
the normal rejoin/backoff logic and rejoin as quickly as possible.

This avoids getting into a situation where users need to manually
hard-reset their switches after a power failure.

Signed-off-by: Jeff Zigler <zigboi@nrfhome.com>