On `make dk` builds, enable shell and temporarily disable trace
The interactive shell lets the user examine Zigbee state and send
commands over the UART, which can be useful for testing.

Trace was previously enabled, but at one point an SDK update broke it
and now zigbee-avc gets stuck in an assertion loop if it's enabled.
This will need to be debugged.

Signed-off-by: Jeff Zigler <zigboi@nrfhome.com>